FIX: file restructuring
